backend/requirements.txt
aerich==0.8.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncclick==8.3.0.7
bcrypt==4.0.1
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dictdiffer==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.2
h11==0.16.0
idna==3.11
iso8601==1.1.0
marshmallow==3.20.1
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.27.1
